requirements.txt
asttokens==3.0.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_67mh6i36ci/croot/certifi_1752653721535/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools==4.59.0
fsspec==2024.6.1
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_resources==6.5.2
ipython==8.18.1
jedi==0.19.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
kiwisolver==1.4.7
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.9.4
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp39-cp39-win_amd64.whl#sha256=afd4c53d4b548db7b600782fd4368ad419de0533ac199a031101a0b99df7eaf7
packaging==25.0
pandas==2.3.2
parso==0.8.4
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
six==1.17.0
stack-data==0.6.3
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchtext==0.18.0
torchvision==0.22.0+cu128
tqdm==4.67.1
traitlets==5.14.3
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
wcwidth==0.2.13
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
zipp==3.23.0
